Fix EventDrops default export without module bundler
