+++ src/withinRange.js
@@ -1,0 +1,7 @@
+export const withinRange = (date, dateBounds) => {
+    const startingDate = Math.min(...dateBounds);
+    const endingDate = Math.max(...dateBounds);
+
+    // @TODO: remove the `new Date()` constructor in the next major version: we need to force it at configuration level.
+    return new Date(date) >= startingDate && new Date(date) <= endingDate;
+};
